REFACTOR: segregated the middlewares
