+++ internal/middlewares/authorization_middleware.go
@@ -1,0 +1,32 @@
+package middlewares
+
+import (
+	"net/http"
+	"strings"
+)
+
+const AuthToken = "GSLC-123-0R" //Dummy token set for now
+
+// AuthorizationMiddleware checks if the request contains the correct token in the header.
+func AuthorizationMiddleware(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		authHeader := r.Header.Get("Authorization")
+
+		// Check if the Authorization header starts with "Bearer "
+		if !strings.HasPrefix(authHeader, "Bearer ") {
+			http.Error(w, "Unauthorized", http.StatusUnauthorized)
+			return
+		}
+
+		// Extract the token from the header
+		token := strings.TrimPrefix(authHeader, "Bearer ")
+
+		if token != AuthToken {
+			http.Error(w, "Unauthorized", http.StatusUnauthorized)
+			return
+		}
+
+		// If the token is valid, proceed to the next handler
+		next.ServeHTTP(w, r)
+	})
+}
