+++ internal/middlewares/recovery_middleware.go
@@ -1,0 +1,15 @@
+package middlewares
+
+import "net/http"
+
+// RecoveryMiddleware recovers from any panics and writes a 500 error.
+func RecoveryMiddleware(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		defer func() {
+			if err := recover(); err != nil {
+				http.Error(w, "Internal Server Error", http.StatusInternalServerError)
+			}
+		}()
+		next.ServeHTTP(w, r)
+	})
+}
